requirements.txt
argcomplete==1.12.3
azure-core==1.19.0
azure-data-tables==12.1.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
debugpy==1.5.0
decorator==5.1.0
entrypoints==0.3
environ==1.0
idna==3.3
importlib-metadata==4.8.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.18.0
jupyter-client==7.0.6
jupyter-core==4.8.1
lxml==4.6.3
matplotlib-inline==0.1.3
msrest==0.6.21
nest-asyncio==1.5.1
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
ppf-datamatrix==0.1.1
prompt-toolkit==3.0.20
Pygments==2.10.0
pypiwin32==223
pysvg-py3==0.2.2.post3
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pywin32==302
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
wxPython==4.1.1
zipp==3.6.0
